Feature/sangshuduo/td 4198 support address sanitizer for develop (#6414)
* [TD-4198]<feature>: support AddressSanitizer on windows.

* fix compile flag.

* fix compile warning for early version visual c++.

* disable memory sanitizer by default.

* fix variable name

* add static lib link.

+++ src/query/src/qAggMain.c
@@ -4687,8 +4687,8 @@ static bool rate_function_setup(SQLFunctionCtx *pCtx) {
   pInfo->correctionValue = 0;
   pInfo->firstKey    = INT64_MIN;
   pInfo->lastKey     = INT64_MIN;
-  pInfo->firstValue  = INT64_MIN;
-  pInfo->lastValue   = INT64_MIN;
+  pInfo->firstValue  = (double) INT64_MIN;
+  pInfo->lastValue   = (double) INT64_MIN;
 
   pInfo->hasResult = 0;
   pInfo->isIRate = (pCtx->functionId == TSDB_FUNC_IRATE);
